Move all points to the geo package
