+++ geo/us.go
@@ -1,0 +1,175 @@
+package geo
+
+import (
+	"github.com/golang/geo/r3"
+	"github.com/golang/geo/s2"
+)
+
+var usPoints = []s2.Point{
+	s2.Point{r3.Vector{-125.003617, 48.495179, 0}},
+	s2.Point{r3.Vector{-123.248444, 48.284028, 0}},
+	s2.Point{r3.Vector{-123.32224, 49.002072, 0}},
+	s2.Point{r3.Vector{-95.15315, 48.999968, 0}},
+	s2.Point{r3.Vector{-94.957389, 49.370194, 0}},
+	s2.Point{r3.Vector{-94.587639, 48.717583, 0}},
+	s2.Point{r3.Vector{-91.5675, 48.043778, 0}},
+	s2.Point{r3.Vector{-88.369862, 48.306063, 0}},
+	s2.Point{r3.Vector{-84.129, 46.5305, 0}},
+	s2.Point{r3.Vector{-82.518611, 45.338639, 0}},
+	s2.Point{r3.Vector{-82.122806, 43.591194, 0}},
+	s2.Point{r3.Vector{-83.149694, 42.041, 0}},
+	s2.Point{r3.Vector{-82.679722, 41.676556, 0}},
+	s2.Point{r3.Vector{-78.935381, 42.828304, 0}},
+	s2.Point{r3.Vector{-79.200667, 43.450472, 0}},
+	s2.Point{r3.Vector{-78.690389, 43.631139, 0}},
+	s2.Point{r3.Vector{-76.796694, 43.631139, 0}},
+	s2.Point{r3.Vector{-74.992806, 44.9775, 0}},
+	s2.Point{r3.Vector{-70.838072, 45.23462, 0}},
+	s2.Point{r3.Vector{-69.22447, 47.45984, 0}},
+	s2.Point{r3.Vector{-67.790198, 47.067225, 0}},
+	s2.Point{r3.Vector{-67.818111, 45.693722, 0}},
+	s2.Point{r3.Vector{-66.885416, 44.794192, 0}},
+	s2.Point{r3.Vector{-67.977722, 43.799501, 0}},
+	s2.Point{r3.Vector{-70.230059, 43.212408, 0}},
+	s2.Point{r3.Vector{-70.47579, 42.301795, 0}},
+	s2.Point{r3.Vector{-69.742916, 42.035449, 0}},
+	s2.Point{r3.Vector{-69.793525, 41.09989, 0}},
+	s2.Point{r3.Vector{-71.149742, 41.250473, 0}},
+	s2.Point{r3.Vector{-73.710867, 40.382469, 0}},
+	s2.Point{r3.Vector{-75.739254, 36.923369, 0}},
+	s2.Point{r3.Vector{-75.325703, 35.10289, 0}},
+	s2.Point{r3.Vector{-78.633116, 33.605186, 0}},
+	s2.Point{r3.Vector{-80.888353, 31.621939, 0}},
+	s2.Point{r3.Vector{-81.21015, 30.584277, 0}},
+	s2.Point{r3.Vector{-79.827577, 26.902276, 0}},
+	s2.Point{r3.Vector{-79.929182, 25.458321, 0}},
+	s2.Point{r3.Vector{-80.245395, 24.85419, 0}},
+	s2.Point{r3.Vector{-81.893254, 24.252523, 0}},
+	s2.Point{r3.Vector{-82.36932, 24.638601, 0}},
+	s2.Point{r3.Vector{-81.379148, 25.371213, 0}},
+	s2.Point{r3.Vector{-82.954626, 27.460737, 0}},
+	s2.Point{r3.Vector{-82.989918, 28.874247, 0}},
+	s2.Point{r3.Vector{-84.054988, 29.86277, 0}},
+	s2.Point{r3.Vector{-85.05918, 29.386094, 0}},
+	s2.Point{r3.Vector{-86.366191, 30.170999, 0}},
+	s2.Point{r3.Vector{-88.601423, 30.01999, 0}},
+	s2.Point{r3.Vector{-88.771509, 29.052116, 0}},
+	s2.Point{r3.Vector{-89.417794, 28.704188, 0}},
+	s2.Point{r3.Vector{-93.991435, 29.478763, 0}},
+	s2.Point{r3.Vector{-96.775786, 27.840658, 0}},
+	s2.Point{r3.Vector{-97.152203, 27.095205, 0}},
+	s2.Point{r3.Vector{-96.92403, 25.975653, 0}},
+	s2.Point{r3.Vector{-97.40561, 25.83764, 0}},
+	s2.Point{r3.Vector{-99.08453, 26.397977, 0}},
+	s2.Point{r3.Vector{-99.5137, 27.56953, 0}},
+	s2.Point{r3.Vector{-101.40112, 29.7712, 0}},
+	s2.Point{r3.Vector{-102.67399, 29.74463, 0}},
+	s2.Point{r3.Vector{-103.28593, 28.97678, 0}},
+	s2.Point{r3.Vector{-104.50828, 29.63249, 0}},
+	s2.Point{r3.Vector{-104.92263, 30.60366, 0}},
+	s2.Point{r3.Vector{-106.45068, 31.76423, 0}},
+	s2.Point{r3.Vector{-111.074825, 31.332239, 0}},
+	s2.Point{r3.Vector{-114.719961, 32.718655, 0}},
+	s2.Point{r3.Vector{-117.46369, 32.589368, 0}},
+	s2.Point{r3.Vector{-117.55455, 33.048611, 0}},
+	s2.Point{r3.Vector{-118.153743, 33.497587, 0}},
+	s2.Point{r3.Vector{-118.281229, 32.625282, 0}},
+	s2.Point{r3.Vector{-118.816446, 33.378432, 0}},
+	s2.Point{r3.Vector{-119.67952, 33.066535, 0}},
+	s2.Point{r3.Vector{-119.855544, 33.511563, 0}},
+	s2.Point{r3.Vector{-119.268588, 33.382705, 0}},
+	s2.Point{r3.Vector{-118.659657, 33.83058, 0}},
+	s2.Point{r3.Vector{-120.644969, 33.907096, 0}},
+	s2.Point{r3.Vector{-120.899533, 34.989218, 0}},
+	s2.Point{r3.Vector{-122.776681, 37.603026, 0}},
+	s2.Point{r3.Vector{-123.32823, 37.673575, 0}},
+	s2.Point{r3.Vector{-123.21929, 38.135944, 0}},
+	s2.Point{r3.Vector{-124.722484, 40.396588, 0}},
+	s2.Point{r3.Vector{-124.360964, 41.486153, 0}},
+	s2.Point{r3.Vector{-124.90688, 42.782148, 0}},
+	s2.Point{r3.Vector{-124.233776, 46.069884, 0}},
+	s2.Point{r3.Vector{-125.003617, 48.495179, 0}},
+}
+
+var alaskaPoints = []s2.Point{
+	s2.Point{r3.Vector{-170.786271, 52.901463, 0}},
+	s2.Point{r3.Vector{-169.178205, 53.128878, 0}},
+	s2.Point{r3.Vector{-168.172791, 54.142362, 0}},
+	s2.Point{r3.Vector{-167.520811, 53.668593, 0}},
+	s2.Point{r3.Vector{-167.166175, 54.149741, 0}},
+	s2.Point{r3.Vector{-165.236204, 54.460239, 0}},
+	s2.Point{r3.Vector{-164.710908, 55.085786, 0}},
+	s2.Point{r3.Vector{-160.768101, 56.244527, 0}},
+	s2.Point{r3.Vector{-158.058169, 57.763424, 0}},
+	s2.Point{r3.Vector{-158.179769, 58.398398, 0}},
+	s2.Point{r3.Vector{-162.378256, 58.4998, 0}},
+	s2.Point{r3.Vector{-162.248738, 59.532214, 0}},
+	s2.Point{r3.Vector{-162.615427, 59.683463, 0}},
+	s2.Point{r3.Vector{-164.041785, 59.491581, 0}},
+	s2.Point{r3.Vector{-165.11255, 60.014776, 0}},
+	s2.Point{r3.Vector{-166.184229, 59.558143, 0}},
+	s2.Point{r3.Vector{-167.637104, 59.959304, 0}},
+	s2.Point{r3.Vector{-167.668268, 60.383862, 0}},
+	s2.Point{r3.Vector{-165.443073, 60.798708, 0}},
+	s2.Point{r3.Vector{-166.577994, 61.628736, 0}},
+	s2.Point{r3.Vector{-165.311704, 62.809629, 0}},
+	s2.Point{r3.Vector{-165.459726, 63.143877, 0}},
+	s2.Point{r3.Vector{-161.232883, 63.816331, 0}},
+	s2.Point{r3.Vector{-162.167463, 64.352259, 0}},
+	s2.Point{r3.Vector{-166.396405, 64.311129, 0}},
+	s2.Point{r3.Vector{-167.401963, 65.218113, 0}},
+	s2.Point{r3.Vector{-169.030549, 65.485543, 0}},
+	s2.Point{r3.Vector{-168.976944, 65.983915, 0}},
+	s2.Point{r3.Vector{-165.001345, 66.729713, 0}},
+	s2.Point{r3.Vector{-162.378703, 66.457652, 0}},
+	s2.Point{r3.Vector{-167.326367, 68.278422, 0}},
+	s2.Point{r3.Vector{-166.648948, 69.013686, 0}},
+	s2.Point{r3.Vector{-164.269546, 69.169597, 0}},
+	s2.Point{r3.Vector{-162.667968, 70.371867, 0}},
+	s2.Point{r3.Vector{-156.602501, 71.599664, 0}},
+	s2.Point{r3.Vector{-140.735184, 69.844169, 0}},
+	s2.Point{r3.Vector{-141.00198, 60.306369, 0}},
+	s2.Point{r3.Vector{-139.074819, 60.352505, 0}},
+	s2.Point{r3.Vector{-137.526114, 58.906372, 0}},
+	s2.Point{r3.Vector{-135.479161, 59.798041, 0}},
+	s2.Point{r3.Vector{-133.377303, 58.430548, 0}},
+	s2.Point{r3.Vector{-131.835327, 56.599203, 0}},
+	s2.Point{r3.Vector{-130.003485, 56.008075, 0}},
+	s2.Point{r3.Vector{-129.974167, 55.281919, 0}},
+	s2.Point{r3.Vector{-130.615673, 54.707474, 0}},
+	s2.Point{r3.Vector{-133.716097, 54.611912, 0}},
+	s2.Point{r3.Vector{-134.123681, 55.653279, 0}},
+	s2.Point{r3.Vector{-134.849133, 55.755657, 0}},
+	s2.Point{r3.Vector{-136.938941, 58.08123, 0}},
+	s2.Point{r3.Vector{-138.458509, 58.893917, 0}},
+	s2.Point{r3.Vector{-142.315846, 59.874249, 0}},
+	s2.Point{r3.Vector{-144.701623, 59.588205, 0}},
+	s2.Point{r3.Vector{-144.821676, 60.015006, 0}},
+	s2.Point{r3.Vector{-145.971396, 60.191974, 0}},
+	s2.Point{r3.Vector{-151.550375, 58.917044, 0}},
+	s2.Point{r3.Vector{-151.918326, 58.700297, 0}},
+	s2.Point{r3.Vector{-151.38742, 58.264993, 0}},
+	s2.Point{r3.Vector{-151.936447, 57.979034, 0}},
+	s2.Point{r3.Vector{-151.90063, 57.308509, 0}},
+	s2.Point{r3.Vector{-153.758168, 56.346107, 0}},
+	s2.Point{r3.Vector{-155.006876, 56.27576, 0}},
+	s2.Point{r3.Vector{-154.653447, 56.820702, 0}},
+	s2.Point{r3.Vector{-155.151209, 57.4368, 0}},
+	s2.Point{r3.Vector{-155.673394, 57.359964, 0}},
+	s2.Point{r3.Vector{-156.609851, 56.615434, 0}},
+	s2.Point{r3.Vector{-156.369729, 55.891094, 0}},
+	s2.Point{r3.Vector{-157.348742, 56.33107, 0}},
+	s2.Point{r3.Vector{-159.231163, 55.548742, 0}},
+	s2.Point{r3.Vector{-158.835098, 54.913615, 0}},
+	s2.Point{r3.Vector{-159.350874, 54.593299, 0}},
+	s2.Point{r3.Vector{-161.089242, 54.971307, 0}},
+	s2.Point{r3.Vector{-162.427977, 54.128008, 0}},
+	s2.Point{r3.Vector{-163.99195, 54.440292, 0}},
+	s2.Point{r3.Vector{-168.29506, 52.807083, 0}},
+	s2.Point{r3.Vector{-171.595163, 52.392697, 0}},
+	s2.Point{r3.Vector{-170.786271, 52.901463, 0}},
+}
+
+var US = &City{
+	points: usPoints,
+}
